text input control test base
