+++ modules/javafx.controls/src/test/java/test/javafx/scene/control/behavior/TextInputControlTestBase.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package test.javafx.scene.control.behavior;
+
+import javafx.scene.control.IndexRange;
+import javafx.scene.control.TextInputControl;
+import org.junit.jupiter.api.Assertions;
+
+/**
+ * Base class for testing behaviors based on TextInputControlBehavior.
+ */
+public class TextInputControlTestBase<T extends TextInputControl> extends BehaviorTestBase<T> {
+    public TextInputControlTestBase() {
+    }
+
+    protected Runnable setText(String text) {
+        return () -> {
+            control.setText(text);
+        };
+    }
+
+    protected Runnable checkText(String expected) {
+        return () -> {
+            String v = control.getText();
+            Assertions.assertEquals(expected, v, errorMessage());
+        };
+    }
+
+    protected Runnable checkSelection(int start, int end) {
+        return () -> {
+            IndexRange v = control.getSelection();
+            IndexRange expected = new IndexRange(start, end);
+            Assertions.assertEquals(expected, v, errorMessage());
+        };
+    }
+}
